# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "fluentlo API",
        "description": "Laravel 12 + Sanctum + Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/admin/onboarding-screens": {
            "get": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Get all onboarding screens (Admin)",
                "operationId": "adminGetOnboardingScreens",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of onboarding screens (admin)"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Create new onboarding screen (Admin)",
                "operationId": "adminStoreOnboardingScreen",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "image",
                                    "background_color",
                                    "text_color",
                                    "button_color",
                                    "order_index"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "subtitle": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "background_color": {
                                        "type": "string",
                                        "pattern": "^#[0-9A-Fa-f]{6}$"
                                    },
                                    "text_color": {
                                        "type": "string",
                                        "pattern": "^#[0-9A-Fa-f]{6}$"
                                    },
                                    "button_color": {
                                        "type": "string",
                                        "pattern": "^#[0-9A-Fa-f]{6}$"
                                    },
                                    "order_index": {
                                        "type": "integer",
                                        "minimum": 0
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Onboarding screen created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/admin/onboarding-screens/{id}": {
            "get": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Get specific onboarding screen (Admin)",
                "operationId": "adminShowOnboardingScreen",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Onboarding screen details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Update onboarding screen (Admin)",
                "operationId": "adminUpdateOnboardingScreen",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "subtitle": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "background_color": {
                                        "type": "string",
                                        "pattern": "^#[0-9A-Fa-f]{6}$"
                                    },
                                    "text_color": {
                                        "type": "string",
                                        "pattern": "^#[0-9A-Fa-f]{6}$"
                                    },
                                    "button_color": {
                                        "type": "string",
                                        "pattern": "^#[0-9A-Fa-f]{6}$"
                                    },
                                    "order_index": {
                                        "type": "integer",
                                        "minimum": 0
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Delete onboarding screen (Admin)",
                "operationId": "adminDeleteOnboardingScreen",
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/admin/onboarding-screens/{id}/toggle-status": {
            "post": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Toggle onboarding screen active status (Admin)",
                "operationId": "adminToggleOnboardingScreenStatus",
                "responses": {
                    "200": {
                        "description": "Status toggled successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/admin/onboarding-screens/reorder": {
            "post": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Reorder onboarding screens (Admin)",
                "operationId": "adminReorderOnboardingScreens",
                "responses": {
                    "200": {
                        "description": "Screens reordered successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}